pyproject.toml
[tool.poetry]
name = "ManageBooks"
version = "0.1.0"
description = ""
authors = ["Michael233ctrl <mikhail.bessarabenko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.37"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.9.1"}
python-multipart = "^0.0.5"
pytest = "^7.1.2"
requests = "^2.28.0"
black = "^22.6.0"
asyncpg = "^0.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
